test: Avoid `virsh iface-list`
This depends on the deprecated `virtinterfaced` which we are getting rid
of. This code was both unnecessarily complicated, and also too
arbitrary: It just took what interface appeared in `virsh iface-list`
first.

Change this to be more realistic: The test creates a new network which
NATs to a host device, so in most practical cases that should be the one
having the default route. Rename the function and add a comment to
clarify this.